requirements.txt
torch>=1.10.0
torchvision>=0.11.0
streamlit>=1.10.0
numpy>=1.20.0
matplotlib>=3.4.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
pillow>=8.3.0 